simplify abstractions
combine filestore and configfile
