+++ pkg/cmd/agent/agent.go
@@ -151,10 +151,7 @@ func Execute(opts *AgentOpts) error {
 	}
 
 	// Create ConfigFileManager for synchronization of shared configuration
-	confManager := config.NewConfigFileManager(&config.ConfigFileManagerOpts{
-		BucketStoreConfig: env.GetKubecostConfigBucket(),
-		LocalConfigPath:   "/",
-	})
+	confManager := config.NewConfigFileManager()
 
 	cloudProviderKey := env.GetCloudProviderAPIKey()
 	cloudProvider, err := cloud.NewProvider(clusterCache, cloudProviderKey, confManager)
